+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Change log
 
-## DEV
+## 4.4.24 (2018-09-05)
 
  * Ignore the "uncached" insert tag flag in the unknown insert tags (see #48).
  * Make the ID of the subscription modules unique (see #40).
+ * Use the correct table when handling root nodes in the picker (see #44).
 
 ## 4.4.23 (2018-08-28)